Update php version

+++ src/RentJeeves/CheckoutBundle/Tests/Functional/AciCollectPayCase.php
@@ -135,10 +135,10 @@ class AciCollectPayCase extends BaseTestCase
         $this->fillForm(
             $form,
             [
-                'rentjeeves_checkoutbundle_paymenttype_amount' => '1000',
+                'rentjeeves_checkoutbundle_paymenttype_paidFor' => $this->paidForStringForCreate,
                 'rentjeeves_checkoutbundle_paymenttype_type' => PaymentTypeEnum::ONE_TIME,
                 'rentjeeves_checkoutbundle_paymenttype_start_date' => (new \DateTime('+2 day'))->format('n/j/Y'),
-                'rentjeeves_checkoutbundle_paymenttype_paidFor' => $this->paidForStringForCreate,
+                'rentjeeves_checkoutbundle_paymenttype_amount' => '1000',
             ]
         );
 
@@ -165,6 +165,8 @@ class AciCollectPayCase extends BaseTestCase
 
         $this->assertNotEmpty($profile = $this->contractForCreate->getTenant()->getAciCollectPayProfile());
 
+        $this->setOldProfileId(md5($this->contractForCreate->getTenant()->getId()), $profile->getProfileId());
+
         $merchantName = $this->contractForCreate
             ->getGroup()->getRentDepositAccountForCurrentPaymentProcessor()->getMerchantName();
 
@@ -187,19 +189,21 @@ class AciCollectPayCase extends BaseTestCase
             "jQuery('#rentjeeves_checkoutbundle_paymenttype_amount:visible').length"
         );
 
-        $this->assertNotEmpty($forms = $this->page->findAll('css', 'form'));
-        $this->assertCount(4, $forms);
+        $form = $this->page->find('css', '#rentjeeves_checkoutbundle_paymenttype');
 
-        $forms[0]->find('css', 'select[name="rentjeeves_checkoutbundle_paymenttype[type]"]')->selectOption(
-            PaymentTypeEnum::ONE_TIME
+        $this->fillForm(
+            $form,
+            [
+                'rentjeeves_checkoutbundle_paymenttype_paidFor' => $this->paidForStringForUpdate,
+                'rentjeeves_checkoutbundle_paymenttype_type' => PaymentTypeEnum::ONE_TIME,
+                'rentjeeves_checkoutbundle_paymenttype_start_date' => (new \DateTime('+2 day'))->format('n/j/Y'),
+                'rentjeeves_checkoutbundle_paymenttype_amount' => '1000',
+            ]
         );
-        $forms[0]->find('css', 'select[name="rentjeeves_checkoutbundle_paymenttype[paidFor]"]')->setValue(
-            $this->paidForStringForUpdate
-        );
-        $forms[0]->find('css', 'input[name="rentjeeves_checkoutbundle_paymenttype[amount]"]')->setValue('1000');
-        $forms[0]->find('css', 'input[name="rentjeeves_checkoutbundle_paymenttype[start_date]"]')->setValue(
-            (new \DateTime('+2 day'))->format('n/j/Y')
-        );
+
+        $this->getDomElement('table.ui-datepicker-calendar a.ui-state-active')->click();
+
+        $this->session->wait(100);
         $this->page->pressButton('pay_popup.step.next');
 
         $this->session->wait(
@@ -217,13 +221,6 @@ class AciCollectPayCase extends BaseTestCase
 
         $this->getEntityManager()->refresh($this->contractForUpdate);
         $this->assertNotEmpty($this->contractForUpdate->getTenant()->getAciCollectPayProfileId());
-
-        $this->setOldProfileId(
-            md5($this->contractForCreate->getTenant()->getId()),
-            $this->contractForCreate->getTenant()->getAciCollectPayProfileId()
-        );
-
-        $this->deleteProfile($this->contractForCreate->getTenant()->getAciCollectPayProfileId());
     }
 
     /**
@@ -359,10 +356,7 @@ class AciCollectPayCase extends BaseTestCase
         );
         $this->page->pressButton('pay_popup.step.next');
 
-        $this->session->wait(
-            $this->timeout,
-            "!jQuery('#id-source-step').is(':visible')"
-        );
+        $this->session->wait($this->timeout, "!$('.overlay').is(':visible')");
 
         $this->assertNotEmpty($errors = $this->page->findAll('css', '.attention-box li'));
         $this->assertCount(2, $errors);
@@ -404,10 +398,10 @@ class AciCollectPayCase extends BaseTestCase
         $this->fillForm(
             $form,
             [
-                'rentjeeves_checkoutbundle_paymenttype_amount' => '1000',
+                'rentjeeves_checkoutbundle_paymenttype_paidFor' => $this->paidForStringForCreate,
                 'rentjeeves_checkoutbundle_paymenttype_type' => PaymentTypeEnum::ONE_TIME,
                 'rentjeeves_checkoutbundle_paymenttype_start_date' => (new \DateTime('+2 day'))->format('n/j/Y'),
-                'rentjeeves_checkoutbundle_paymenttype_paidFor' => $this->paidForStringForCreate,
+                'rentjeeves_checkoutbundle_paymenttype_amount' => '1000',
             ]
         );
 
@@ -437,6 +431,11 @@ class AciCollectPayCase extends BaseTestCase
         $this->getEntityManager()->refresh($this->contractForCreate);
         $this->getEntityManager()->refresh($this->contractForCreate->getTenant());
 
+        $this->setOldProfileId(
+            md5($this->contractForCreate->getTenant()->getId()),
+            $this->contractForCreate->getTenant()->getAciCollectPayProfileId()
+        );
+
         $this->assertNotEmpty($profile = $this->contractForCreate->getTenant()->getAciCollectPayProfile());
 
         $merchantName = $this->contractForCreate
@@ -460,11 +459,6 @@ class AciCollectPayCase extends BaseTestCase
             $paymentAccount->getDebitType(),
             'Created Payment Account should be have debit_type "debit"'
         );
-        $this->setOldProfileId(
-            md5($this->contractForCreate->getTenant()->getId()),
-            $this->contractForCreate->getTenant()->getAciCollectPayProfileId()
-        );
-        $this->deleteProfile($this->contractForCreate->getTenant()->getAciCollectPayProfileId());
     }
 
     /**
@@ -509,10 +503,10 @@ class AciCollectPayCase extends BaseTestCase
         $this->fillForm(
             $form,
             [
-                'rentjeeves_checkoutbundle_paymenttype_amount' => '1000',
+                'rentjeeves_checkoutbundle_paymenttype_paidFor' => $this->paidForStringForCreate,
                 'rentjeeves_checkoutbundle_paymenttype_type' => PaymentTypeEnum::ONE_TIME,
                 'rentjeeves_checkoutbundle_paymenttype_start_date' => (new \DateTime('+2 day'))->format('n/j/Y'),
-                'rentjeeves_checkoutbundle_paymenttype_paidFor' => $this->paidForStringForCreate,
+                'rentjeeves_checkoutbundle_paymenttype_amount' => '1000',
             ]
         );
 
@@ -541,6 +535,11 @@ class AciCollectPayCase extends BaseTestCase
         $this->getEntityManager()->refresh($this->contractForCreate);
         $this->getEntityManager()->refresh($this->contractForCreate->getTenant());
 
+        $this->setOldProfileId(
+            md5($this->contractForCreate->getTenant()->getId()),
+            $this->contractForCreate->getTenant()->getAciCollectPayProfileId()
+        );
+
         $this->assertNotEmpty(
             $profile = $this->contractForCreate->getTenant()->getAciCollectPayProfile(),
             'Profile should be created'
@@ -567,10 +566,12 @@ class AciCollectPayCase extends BaseTestCase
             $paymentAccount->getDebitType(),
             'Created Payment Account should be have debit_type "signature_non_exempt"'
         );
-        $this->setOldProfileId(
-            md5($this->contractForCreate->getTenant()->getId()),
-            $this->contractForCreate->getTenant()->getAciCollectPayProfileId()
-        );
-        $this->deleteProfile($this->contractForCreate->getTenant()->getAciCollectPayProfileId());
+    }
+
+    protected function tearDown()
+    {
+        if ($this->contractForCreate && $this->contractForCreate->getTenant()->getAciCollectPayProfileId()) {
+            $this->deleteProfile($this->contractForCreate->getTenant()->getAciCollectPayProfileId());
+        }
     }
 }
